Mobile Grid Based Parking Assistant - Android
This Android application is a simple and elegant solution to several
problems in a parking system. The problems such as high usage of paper
and mobile token printers, time wastage during search of empty slots.
With the concept of networking and communication among smart phones, the
closest empty slot in an area can be found using a single tap on the
screen.